DOCS Tweak description of threadPool.on
Previous description:
"...is like `thread.on()`, registers listeners for events from any of the threads in the pool."

Problem:
This makes it sound like the threads will emit events and that the
threadPool will listen for them.

The actual behavior is to call thread.on() for each thread in the pool.

Solution:
Update the README to reflect this.